Add make_trackpad_toggle, remove trackpad_toggle
The id of the device in the script is subject to change on different
devices, and likely different installs. There's nothing specifically
identifying the device as a trackpad in the xinput output.

This adds a script that will prompt the user for the id after
displaying the xinput, and writes the toggle script.